#505 replace deprecated supervisor.spec
