+++ apps/aecore/test/aecore_channels_test.exs
@@ -110,8 +110,8 @@ defmodule AecoreChannelTest do
   end
 
   @tag :channels
-  @tag timeout: 340_000
-  test "create channel, transfer twice, slash with old, slash with corrent and settle", ctx do
+  @tag timeout: 500_000
+  test "create channel, transfer twice, slash with old, slash with correct and settle", ctx do
     id = create_channel(ctx)
 
     # Can't transfer more then reserve allows
